+++ src/deprecated/deprecated/cache_utils.py
