+++ src/deprecated/deprecated_generate_text.py
